update providers in pages and move session provider
