+++ src/app/admin/layout.tsx
@@ -1,8 +1,8 @@
 import "../globals.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import { Header, Provider } from "@/components";
-import { SessionProvider } from "@/components/atoms/SessionProvider";
+import { Header, ThemeProvider } from "@/components";
+import { SessionProvider } from "@/components/atoms/Providers/SessionProvider/SessionProvider";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -22,7 +22,7 @@ export default function RootLayout({
       <head></head>
       <body className={"bg-white text-black dark:bg-black dark:text-white"}>
         <SessionProvider>
-          <Provider>
+          <ThemeProvider>
             <div className="min-h-screen flex justify-center items-center">
               <div className="bg-white dark:bg-zinc-900 w-full sm:w-full md:w-11/12 lg:w-11/12 xl:w-4/5 max-w-screen-xl p-4 transition-width">
                 <div className="min-h-screen">
@@ -31,7 +31,7 @@ export default function RootLayout({
                 </div>
               </div>
             </div>
-          </Provider>
+          </ThemeProvider>
         </SessionProvider>
       </body>
     </html>
